+++ src/ext_impls/mod.rs
@@ -1,0 +1,20 @@
+#[cfg(feature = "alloc")]
+mod impl_alloc;
+
+#[cfg(feature = "const-default")]
+mod impl_const_default;
+
+#[cfg(feature = "serde")]
+mod impl_serde;
+
+#[cfg(feature = "zeroize")]
+mod impl_zeroize;
+
+#[cfg(feature = "subtle")]
+mod impl_subtle;
+
+#[cfg(feature = "arbitrary")]
+mod impl_arbitrary;
+
+#[cfg(feature = "bytemuck")]
+mod impl_bytemuck;